[INTERPRETER] * BUG: Another fix in the bits manipulation routines.
git-svn-id: svn://localhost/gambas/branches/2.0@1519 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_subr_test.c
@@ -166,7 +166,7 @@ __END:
   if (type == T_LONG)
     RETURN->_long.value = val;
   else
-    RETURN->_integer.value = val & mask;
+    RETURN->_integer.value = (int)val & mask;
 
   SUBR_LEAVE();
 }
